feat(error handling): refactor to errorHandler and refactor chip for readability
