+++ src/components/Chip/Chip.tsx
@@ -1,66 +1,13 @@
-import { ClickEvent } from 'hooks/useLoading';
 import React from 'react';
-import { flatColors } from 'theme/palette';
-import { generateTestDataId, propsHandler } from 'utils/helpers';
-import { TestProps } from 'utils/types';
+import { generateTestDataId, errorHandler } from 'utils/helpers';
 
 import { BASE_SHADE } from '../../theme/palette';
-import { DivProps } from '../../utils/common';
 import { chipStyle, closeIconWrapperStyle } from './Chip.style';
+import { ChipProps } from './Chip.types';
 import Badge from './components/Badge';
+import { defaultProps, errors } from './utils';
 import Avatar from 'components/Avatar';
 import Icon from 'components/Icon';
-
-type ChipProps = Props & TestProps & DivProps;
-
-const READ_ONLY = 'read-only' as const;
-const INTERACTIVE = 'interactive' as const;
-
-const styleType = [READ_ONLY, INTERACTIVE] as const;
-
-const defaultProps = {
-  disabled: false,
-  styleType: READ_ONLY as typeof styleType[number],
-  dataTestId: '',
-};
-
-export type Props = {
-  /**
-   * Determines whether the chip should be read-only or interactive.
-   * @default read-only
-   */
-  styleType?: typeof styleType[number];
-  /** Defines the fill color of the component, if filled. */
-  fill?: typeof flatColors[number];
-  /** An optional thumbnail to show to the left. */
-  thumbnail?: { src?: string; name?: string };
-  /** Boolean defining if the chip is selected. */
-  isSelected?: boolean;
-  /** Boolean defining if the check icon is shown. */
-  isChecked?: boolean;
-  /** Defines the number value of the badge */
-  badgeNumber?: number;
-  /** Callback function for onClick. */
-  onClick?: (event: ClickEvent) => void;
-  /** A callback that is being triggered when type is interactive and you press the X icon. */
-  onClear?: () => void;
-  /** Boolean defining if the chip is disabled. Interactive only. */
-  disabled?: boolean;
-};
-
-const errors = [
-  {
-    condition: ({ styleType, isSelected, isChecked, badgeNumber, disabled }: Props) =>
-      Boolean(styleType === READ_ONLY && (isSelected || isChecked || badgeNumber || disabled)),
-    message:
-      'The properties isSelected, isChecked, badgeNumber and disabled are only for Interactive style type Chips.',
-  },
-  {
-    condition: ({ styleType, thumbnail }: Props) => Boolean(styleType === INTERACTIVE && thumbnail),
-    message:
-      'The properties isSelected, isChecked, badgeNumber and disabled are only for Interactive style type Chips.',
-  },
-];
 
 const Chip = React.forwardRef<HTMLButtonElement, ChipProps>((props, ref) => {
   const {
@@ -75,7 +22,7 @@ const Chip = React.forwardRef<HTMLButtonElement, ChipProps>((props, ref) => {
     dataTestId,
     children,
     badgeNumber,
-  } = propsHandler<ChipProps>(errors, {
+  } = errorHandler<ChipProps>(errors, {
     ...defaultProps,
     ...props,
   });
